Rename and clean up some haku form components
