--- src/main/app/src/pages/haku/HakuForm/TargetGroupSection.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import HaunKohdejoukkoFields from '#/src/components/HaunKohdejoukkoFields';
-
-const TargetGroupSection = ({ name }) => {
-  return <HaunKohdejoukkoFields name={name} />;
-};
-
-export default TargetGroupSection;
